agent/pyproject.toml
[project]
name = "commitment-agent"
version = "0.1.0"
description = "Claude agent for end-to-end commitment evaluation"
requires-python = ">=3.11"
dependencies = [
    "claude-agent-sdk",
    "psycopg2-binary>=2.9.9",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
